+++ src/main/java/com/vedruna/projectmgmt/services/TechnologyServiceImpl.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.vedruna.projectmgmt.dto.CreateTechnologyDTO;
+import com.vedruna.projectmgmt.exceptions.TechnologyAlreadyExists;
+import com.vedruna.projectmgmt.exceptions.TechnologyNotFound;
 import com.vedruna.projectmgmt.persistance.model.Technology;
 import com.vedruna.projectmgmt.persistance.repository.TechnologyRepositoryI;
 
@@ -20,25 +22,41 @@ public class TechnologyServiceImpl implements TechnologyServiceI {
     private TechnologyRepositoryI technologyRepo;
     @Autowired
     private static final Logger log = LoggerFactory.getLogger(TechnologyServiceImpl.class);
+    
+
+    /**
+     * Guardar tecnología.
+     *
+     * Convierte el objeto CreateTechnologyDTO en un objeto Technology y lo
+     * guarda en la base de datos.
+     *
+     * @param technology El objeto CreateTechnologyDTO que se va a guardar.
+     */
     @Override
-    public void saveTechnology(CreateTechnologyDTO tehnology) {
-       Technology tech = new Technology();
-       try{
-        tech.setTechName(tehnology.getTechName());
+    public void saveTechnology(CreateTechnologyDTO technology) {
+        if(technologyRepo.existsByTechName(technology.getTechName())){
+            throw new TechnologyAlreadyExists(technology.getTechId());
+        }
+        Technology tech = new Technology();
+        tech.setTechName(technology.getTechName());
         tech.setProjects(Collections.emptyList());
         technologyRepo.save(tech);
-       }catch(Exception e){
-           log.error(e.getMessage());
-       }
     }
 
-    @Override
-    public void deleteTechnology(Integer id) {
-        try{
-        technologyRepo.deleteById(id);
-        }catch(Exception e){
-            log.error(e.getMessage());
+    /**
+     * Eliminar tecnología.
+     *
+     * Elimina una tecnología por su id. Comprueba si la tecnología existe
+     * Si no es así, lanza una excepción personalizada
+     *
+     * @param id El id de la tecnología a eliminar.
+    */
+         @Override
+         public void deleteTechnology(Integer id){
+        if(!technologyRepo.existsById(id)){
+            throw new TechnologyNotFound(id);
         }
+        technologyRepo.deleteById(id);
     }
     
 }
